[External] [stdlib] Introduce non owning collection type (#39731)
[External] [stdlib] Introduce non owning collection type

This PR introduces a non owning collection type for contiguous arrays of
`CollectionElement`. With this type, we will be able to tie together the
array types in the standard library, and reduce necessary copies in some
places.

### Motivation

The pointer & length data structure is common across most mainstream
programming languages at this point
([Zig](https://ziglang.org/documentation/master/#Slices),
[Rust](https://doc.rust-lang.org/std/primitive.slice.html),
[C++](https://en.cppreference.com/w/cpp/container/span)). In order to
mirror APIs developers are used to with the performance characteristic
they expect, this type will be required.

Notably, it now allows functions that can work on `List` _or_ `Array`
without an overload:
```mojo
fn copy[
    T: CollectionElement, lifetime: MutLifetime
](owned dst: Span[T, __mlir_attr.`1: i1`, lifetime], src: Span[T, _, _],):
    for i in range(len(src)):
        dst[i] = src[i]

fn main():
    var l = List[Int](1, 2, 3)
    var a = InlineArray[Int, 3](4, 5, 6)
    copy(Span(l), Span(a))
    for i in l:
        print(i[])
```
This code keeps the `List` and `InlineArray` alive long enough for the
copy to work, which is not true for `Buffer` or `DTypePointer`.

This type will be especially necessary for traits such as `Read` and
`Write`, which will want to take a `Span` as the relevant buffer, since
taking a `List` or `Array` is use case dependent for such functions.

Another motivation for this change is to reduce unnecessary copies on
methods such as slicing a `List`, which currently allocates a new list.
Ideally, we would just return a `Span` without making a copy. The same
could be said for `String.as_bytes()`, which does not need to allocate a
`List`.

**Note about naming**

I'm using the name `Span` to mirror C++'s `std::span<T>`. Rust, Go, and
Zig call this type a `slice`, but that term has an overloaded meaning in
Python. It's also nice that it's short as a name. Similarly, Mojo
already has the `Buffer` type. I'm open to other options if `Span` is
not the right fit.

Co-authored-by: Lukas Hermann <lukashermann28@gmail.com>
Closes modularml/mojo#2595
MODULAR_ORIG_COMMIT_REV_ID: 2fb5385d4d2f141150088ebd3e107cebdba761d2