[mojo-lang] Expand `__lifetime_of` to be variadic (#39748)
This PR expands the `__lifetime_of` magic function to take any number of
arguments and returns a union of the lifetimes of the arguments. If
there are no arguments, it returns a `#lit.lifetime` attribute. Any more
than 1, and it returns a lifetime union. If any of the lifetimes are
immutable, the overall lifetime is mutcast'd to immutable.

MODULAR_ORIG_COMMIT_REV_ID: e28fe21dedd8ac45c1abf1eb14b2a9ee9993ef5e