[mojo-lang] Refactor `getRValueType()`, NFC. (#39819)
With auto-deref for `Reference` inbound, we have driving into a
world with more of a difference between BValue/LValue and RValue,
introducing a conversion that happens "after the load".  This means
that the `getRValueType` method will need to be slightly more
complicated in the future, and many of them will need access to
the SharedState in order to reason about the types they're working
with.

Start phasing this in by adding SharedState members to many of the
`getRValueType()` methods, and introduce a new
`LValue::getLValueStorageType()` for clients that actually care
what the LValue is referring to (not the loaded RValue type).

MODULAR_ORIG_COMMIT_REV_ID: 0f8dd4c63d1b8ca7f31116d1a83d28ad950531c9