[mojo-lang] Implementation mechanics for Reference autoderef.
This is more progress towards implementing automatic deference for
`Reference`.  This:

1) Introduces a new `@automatically_dereference` struct decorator which
   signs a type up for autoderef and requires the type to have an
   `__mlir_ref__` member.
2) Adds type checking logic for that.
3) Adds `@automatically_dereference` to the various references in the
   testsuite and stdlib.
4) Switches some testcases from using `Reference(x).value` to use
   `__get_mvalue_as_litref(x)` instead.  The former will autoderef
   in the future and be synonymous for `x.value` so move these test
   cases, to reduce future diffs.

For what it's worth, I tried to do this without a decorator, but that
suffers
from a big problem: it means that we cannot determine what the RValue
type of a value is without body resolving the type of the expression. We
currently are lazy on this, and I don't want to start resolving lots of
more
types just because of references - a decorator allows us to just
signature
resolve the target struct, which we're already doing.

Autoderef is not implemented yet, this is just progress towards it.

MODULAR_ORIG_COMMIT_REV_ID: ab40ebf86f06bb1dd841293fd0197deb88b1ee3d