[LSP] Basic LSP Server proxy (#39620)
This adds a basic LSP Server proxy in typescript. Given the restrictions
imposed by VSCode, I ended up writing a typescript subproject for the
proxy, which is in fact an LSP server implemented in typescript that is
a passthrough around the actual mojo lsp server. This proxy is a
different process from the extension. Who knows why VSCode enforces
that!

This code effectively acts as a middleware between the client and the
Mojo LSP Server, with the following features
- All notifications and requests supported by the Mojo LSP Server are
also supported by the proxy.
- The proxy can replicate the exit status of the mojo-lsp-server binary.
- No smart logic has been added around crash recovery, that'll be added
later.
- A logger is present to show messages in the VSCode output channel.
- There's an example of shared code between the extension and the proxy,
which in this case is just a type definition

<img width="798" alt="Screenshot 2024-05-10 at 12 35 20 AM"
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 3661a304081f37353da9e0fd0ec1e5305a90253c