[mojo-lang] Fix owned variadic packs to be owned by callee. (#39863)
It's important that a owned variadic packs are owned on the callee
side, not the caller side, because we want the elements to be
destroyed ASAP as soon as the pack is last used.  In the testcase
the pack gets destroyed before the 'owned_variadic_pack done' call.

MODULAR_ORIG_COMMIT_REV_ID: 5eeba0e5613c2e5e6c2d130dd6978744689e3e6e