[REPL] Enable symbol index to speed up initialization (#39893)
As part of speeding up initialization, I was playing with symbol tables.
I tried to force not processing symbol tables, but this is not possible
because many core parts of LLDB rely on symbol table processing in
different ways. So the best we can do is to enable the symbol cache and
allow LLDB to do its job without "hacks". This saved a solid 300ms on my
machine.

With the symbol cache enabled, the first time symbol tables are
processed, which is slowish, they are converted into a mmap'able format
and stored on disk. Then, subsequent runs of LLDB can just mmap these
symbol tables without paying processing time.

There are still symbol table lookups due to the different LLDB features
that require it, but they are O(log). As a way to speed this up a bit
further, I'm also disabling a few features that make use of symbol
tables and are optional. They didn't really improve the runtime in a
noticeable way, but better safe than sorry.

MODULAR_ORIG_COMMIT_REV_ID: f1189c68dd207623bd9ac9c02e1d763d9c3007d9