[Kernels] add M=1 int4 kernels for LLM token gen (#39948)
This adds kernels to handle the M=1 vector-matrix multiply path that is
seen during token generation for LLMs.

The existing kernels do an int4->int8 unpack of slices of the packed B
matrix to the stack. These unpacked slices are then used multiple times
across the M dimension of the A matrix. This then avoids needing to do
expensive repacking in the inner kernel and was one of the last
techniques used to speed up the int4 kernels for context encoding.

However, this slows down the M=1 case where there is no reuse of the
unpacked slices, so the unpack step wastes cycles. This change brings
back part of an older version of the kernel that did the int4 unpack
inline with the dot product loop, but now it only needs to handle this
special case.

This helps the most on Graviton3/Mac, but I'm seeing small boosts on
Intel as well. i8mm instruction support doesn't help with the single row
case, so there is a single kernel for Graviton3/Mac.

---------

Co-authored-by: abdul dakkak <adakkak@modular.com>
QUANTIZATION_ORIG_REV_ID: 4b9c3c33c463946c01715acca7c875821e879ea8