[mojo-lang] Lower Resume Functions (#39593)
This patch transforms `coroutine.await` ops in a function marked as
"coro" into a state update + return followed by a block with the code
contained in the region. It also adds a state switch to the front of the
function.

MODULAR_ORIG_COMMIT_REV_ID: 233910c0e183bc027acaae157209f98c0bf506c3