[KGEN] Delete no-lldbinit test (#39969)
This test wasn't actually testing what it meant to be, because the
.lldbinit file wasn't in the same dir as where the test was being run
from. When we fix that it actually doesn't pass since the local
.lldbinit file is checked before our custom lldbinit disabling that is
run. We decided we don't need this since it is otherwise just testing
default lldb behavior.

MODULAR_ORIG_COMMIT_REV_ID: 8d580251751b9d12413aabe68021dd5af6f2998c