[bazel] Avoid lldb-argdumper macOS issues (#39982)
On macOS lldb's `run` command requires lldb-argdumper, which lldb
attempts to discover "near" the lldb library. With bazel we don't
reproduce this structure, so it can't find it and fails to run. We can
temporarily avoid with `process launch` but once we update the submodule
we can drop this for a new env var to specify the path

MODULAR_ORIG_COMMIT_REV_ID: e631aa78fcb338a8a678a243fd15cb93a850164c