[MojoBuild] Build project at specified path (#38895)
Expand the build server protocol supported by `mojo-build-server` to
include the `rootUri` member of `BuildInitializeParams`. The idea here
is that the server will be initialized with the path to a directory
containing a Mojo project manifest, or to a specific Mojo project
manifest file, and use that to build the project. This patch also adds
input option handling to the `mojo build-project` command, to pass along
that path when initializing the build server.

This patch does not implement the actual project building; that will
come in a future commit.

MODULAR_ORIG_COMMIT_REV_ID: e5c4f11fb250ffc656e2fabbc0e2c4a273a855f0