[MojoBuild] Support "buildTarget/compile" method (#39089)
Add the JSON-RPC foundation for sending and responding to the
"buildTarget/compile" method. An upcoming patch will have the server
respond to this method by building a Mojo package based on the root URI
of the workspace it was initialized with. For now, however, the server
doesn't really do anything when sent this message.

MODULAR_ORIG_COMMIT_REV_ID: 2c5a748e4d8277d87787a62958f6bb5f41002539