[mojo-stdlib] Change `__reversed__` to return a `_StridedRange` (NFC) (#40099)
BEGIN_PUBLIC
[stdlib] Change `__reversed__` to return a `_StridedRange` (NFC)

This PR changes the `__reversed__` method on the `ReversibleRange` trait
to reutrn a `_StridedRange` instead of a `_StridedRangeIterator`. This
allows the `__iter__` method to be removed from `_StridedRangeIterator`,
since it didn't make a lot of sense to return an iterator where a range
was required.

This is required deck chair rearranging to support `@parameter for`.

END_PUBILC

MODULAR_ORIG_COMMIT_REV_ID: bb9c9ea1ff161280c75e70233ef5c860368238fa