[KGEN] Disable argument-passthrough with asan (#40111)
When running with asan %repl has env vars in the invocation, and when
running with not they are interpreted as the binary name. It's not super
important that this runs with asan anyways.

MODULAR_ORIG_COMMIT_REV_ID: 7a0aa58d43487d73ae5d49edcc27b5138c4e8a1e