[mojo-stdlib] Move intel_amx test to LinAlg
This test doesn't purely depend on the stdlib, so it makes more sense
with the other LinAlg tests. It was also using the wrong REQUIRES so it
was never running, this updates to make it pass again.

MODULAR_ORIG_COMMIT_REV_ID: dcd34acabcfd64336e1232c1c5c6adee1df7ab69