[mojo-lang] Implement parameter for with a generator (#40138)
This PR implements `@parameter for` by parsing the body into a parameter
closure and then emitting a call to the `parameter_for` generator in the
builtins. This approach has the major drawback that early exits don't
work in the body, leading to confusing errors to the user.

In a follow-up, the approach will be refined to use a KGEN dialect op,
so that early exits will work.

MODULAR_ORIG_COMMIT_REV_ID: fd294f0719eaba4e659052097f52e8858f50f0d2