[LSP] Use byte-encoding when talking to mojo-lsp-server (#40119)
The communication between the proxy and mojo-lsp-server is breaking when there are non-ascii characters present. Fixing this is trivial. We just need to read and writes the JSON protocol packets as bytes and not as regular strings.

A sample file with this issue is test_linalg_matmul_gpu.mojo

MODULAR_ORIG_COMMIT_REV_ID: f7b1d493a81ad6aa68c1ff9196ac948806bc7788