[Stdlib] Extend the float cast to be able to cast to non-f64 types
This allows one to cast the `FloatLiteral` constant tO types other than
`f64`. This enables us to avoid numerics issues that might arise due to
upcasting and then downcasting and instead perform all the operations in
the arbitrary precision domain. Due to weakness in the mojo
parameterization, we have to duplicate a bunch of code, but all of this
is under a param if, so does not cause issues at runtime.

MODULAR_ORIG_COMMIT_REV_ID: 35f6cd11726838375908ab408c2a871172a7dc34