[Tooling] Perform the mojo invocation log asynchronously (#40058)
The REPL is taking 1 second on my machine just to perform synchronous
telemetry for the main mojo invocation event. This might be due to me
being in Italy (not USA), which means that very likely our non-US users
are experiencing the same slowdown.
I was able to reduce this time by making the log happen in a different
thread.
I also had to change the way we invoke LLDB to not use `execv`, which
wouldn't ensure we wait for the log thread to join, but that's a small
price to pay.

MODULAR_ORIG_COMMIT_REV_ID: 8c476a4ccc5bdab9f0776253e7da616abc9355f3