[mojo-lang] Remove one overload of `verifyBindings` (NFC) (#40237)
By making the methods take an optional pointer to `diagEmitter`, we can
remove one overload of the function. This is important, because I
personally find it hard to keep track of each overload and the nuanced
differences in how they are used.

Also renames the overload that actually implements the logic, but which
is not used exposed, to `verifyBindingsImpl`, just to make it easier to
search for.

MODULAR_ORIG_COMMIT_REV_ID: 9b448b35f90e06a32a2cba237a598733f1aa3d57