[mojo-lang] Rework 'def' argument processing
We previously defaulted 'def' arguments to 'owned' when no explicit
arg convention is specified.  This enabled them to be mutable in the
body of the function like Python.  However, this came at a serious
cost: it made all 'def' arguments get copied in the caller side,
which is extremely impractical for expensive-to-copy arguments and
doesn't work for non-copyable types.

This changes Mojo to use a computed lvalue to lazily synthesize the
argument box only if it is mutated. This results in a lot less IR,
doesn't
copy the value unless there is a reason to and generally provides a
better UX.

This fixes MOCO-179.

MODULAR_ORIG_COMMIT_REV_ID: 4a187f53ed160ac4f3db2d5f7f4b2e110a162cea