[mojo-lang] Improve implicit parameter inference failure diagnostics (#40239)
The diagnostic emitter hook for this would try to dig out the argument
and struct parameter that defined the autoparam in order to emit a nice
error. But, this only worked for SB and SR value arguments, because it
directly casted the argument type to a DeclRefType. Instead, properly
dig out the DeclRefType with a walker.

MODULAR_ORIG_COMMIT_REV_ID: 34b257c3b977a9259a20a46196c0ac325bc2ee65