[mojo-lang] Remove `Boundness::Explicit` (NFC) (#40240)
We originally intended to remove implicit partial binding from the
language. This means that in order to partially bind a type or function,
you had to explicitly unbind parameters using `_` or `*_` syntax. This
was enforced on functions only and we didn't manage to enforce this one
types.

This is because when binding a function, we can tell when a
parameter has been explicitly unbound (is an UnboundAttr) and when it is
not (the value is missing entirely from ParamBindings). On types,
however, we always form the initial reference with a list of UnboundAttr
to represent that the type is fully unbound.

Thus, we couldn't actually tell when a type parameter was explicitly
unbound in general. There is currently no alignment on what the right
solution is as well as no urgency to actually make this change. One
approach would be to introduce a version of OverloadSet but for types
(except it only has one ASTDecl, not a list). This way, we can carry
around the unbound type reference and a ParamBindings list.

In the meantime, we can reduce the complexity of the system by remving
`Boundness::Explicit`, and returning to implicit partial binding vs.
fully bound modes. This walks back the requirement that function binding
be explicitly unbound, but this was confusing anyways for two reasons:

1. It makes the binding rules between functions and types different.
   This is still the case in many other ways and we should merge them.
2. Worse, you had to explicitly unbind parameters when forming a PValue
   reference, but not if you immediately call the function. This is a
   huge inconsistency that meant that implicit partial binding of
   functions was effectively the norm.

MODULAR_ORIG_COMMIT_REV_ID: e4c3fae4939a13c40fc93ea0083acaeb4506edca