[Kernels] int4 kernel cleanup: improved A block format (#40244)
This changes the layout of the internally quantized A buffer to simplify
the inner kernels and share a common packing routine for all kernel
types.

In the old scheme, the input matrix A[N][K] was quantized to a buffer
also of shape a_quant[N][K] and a_scale[N][K // 32] (ignoring i8mm for
the moment). The inner kernels needed to be given K to index into this
a_quant/a_scale. This burns a few extra registers for indexing and data
is spread through memory.

The new scheme interleaves the rows of quantized data as was required
for i8mm to remove the need for indexing using K. The innermost kernel
is a chunk of code that no longer cares about the outer shape of the
matmul.

The A buffers are also packed in terms of the K_BATCH_SIZE of 512. The
kernel loop already unpacked int4->int8 weights from B to a stack buffer
in blocks of this size. A is packed with this as well, so that as the
kernel loop iterates across M, the next K_BATCH_SIZE block of A data is
located right after the previous tile of rows.

None of this changed runtime performance.

QUANTIZATION_ORIG_REV_ID: 501ed20be84f575b20926483c966100fef3b451d