[External] [stdlib] Add List constructor from Span (#40246)
[External] [stdlib] Add List constructor from Span

Being able to convert a Span to an owned value is useful, and List makes
the most sense for this type. We could also consider a `to_list()`
method on Span instead if that is a better fit.

Co-authored-by: Lukas Hermann <lukashermann28@gmail.com>
Closes modularml/mojo#2754
MODULAR_ORIG_COMMIT_REV_ID: 5c4bf563fc8d7644b8c0cab316b5376988e25748