[mojo-tooling][EZ] Manually Move Result of `discoverTestsInDirectory()` (#40268)
The copy constructor for `AnyAsyncValueRef` is deleted so on GCC
compilers the compiler will error and say we are using the deleted copy
constructor.

On Clang compilers, the copy will be elided to a move automatically it
seems.

This PR manually calls a move to allow compilation with gcc toolchains
(which appear to be the default on mdcm)

MODULAR_ORIG_COMMIT_REV_ID: fb14318c1c300976ab22396a5a4216929a3124ae