[KGEN] Change the branch folding threshold for simplify CFG on GPUS (#40226)
For GPUs branches are pretty expensive (causing divergence), so we want
to work harder to remove them. This increases the threshold in the CFG
pass to ensure more branches are folded. This optimization is done in
the XLA compiler and is also used by the AMD backend of LLVM.

MODULAR_ORIG_COMMIT_REV_ID: e80240bd399dd118df808033d4164606bb7bec34