[LSP] Fix a crash caused by synthetic arguments (#40412)
It seems that there's a synthetic argument `__result__` somewhere in
`MOGG.mojo` which was causing a crash when producing the list of symbols
for this file. As a fix, I'm just making the argument decl view creator
more robust by double checking the bounds of the list of arguments of
the parent function.

Closes MOTO-486
MODULAR_ORIG_COMMIT_REV_ID: 986128fe67cb2e7826f23c21c4ad288d9b977aa6