[mojo-lang] Introduce `load_from_mem` POC and teach the parser to emit it (#40613)
This is part of enabling `__refitem__` in parameter expressions. This
introduces a new parameter operator that requires the IREvaluator to
resolve (because loading requires a target), and teaches the parser to
emit `__refitem__` in the parameter domain.

MODULAR_ORIG_COMMIT_REV_ID: 49183f19a5d7d9364ba44f21a7c759833cda7299