[KGEN] Canonicalize `kgen.param.if` with a constant condition (#40628)
We can fold away `kgen.param.if` with a constant known condition, but we
have to be careful: it declares a new parameter scope, so we have to
keep the operation around if there are ops in the body. We can at least
reduce IR by replacing the dead block with an `unreachable`.

MODULAR_ORIG_COMMIT_REV_ID: d340412c4f5a6c46ee7bd222c28e5c7e3394506a