[KGEN] Disable the interpreter in parameter simplification (#40661)
This is not correct; it is generating invalid IR and can lead to
crashes.
This is the same problem the parser faces with eager interpretation,
except the point is to reduce IR, so generating rebinds doesn't work.

Leave in the feature because it is used to test `::interpret` hooks, but
we need another strategy for simplifying `apply` expressions before
elaboration.

This should fix the root cause in
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 079f27ac8f55051eadffb1998f9c02b59ba39343