[mojo-lang] Tidy up for statement parsing logic, add testcase. (#40689)
This adds a testcase for a collection with an iterator that uses a
`__next__` returning a reference.  As expected, we automatically load
this value into the declared temporary.  This won't enable mutation
of the collection elements.

We need to discuss how this should work, e.g. do we introduce ref
patterns now?

MODULAR_ORIG_COMMIT_REV_ID: 3caddfbc633f9b3f58577bbb99515b0f8e4445d6