[mojo-lang] Forbid taking references to maybe reg variadics (#40665)
We forbid taking references to arguments that might be register passable
because they don't necessarily have an address. This is because we use
the memory-only form for generality in the frontend, but
post-elaboration, we convert the calling convention to respect the
register-passability of the type.

We want to do the same with variadics and packs, so start enforcing this
requirement.

MODULAR_ORIG_COMMIT_REV_ID: fe00bd73a78741f89ac13f890e3c949a7890fcb2