[mojo-lang] Frame Transformation (#40327)
Calculates the frame struct type by determining the state for each op
in the function. If a value's use state is larger than its definition
state it is added to the frame. The state of a virtual block is the max
of the state of its predecessors. A coroutine.await advances the state
of a virtual block.

MODULAR_ORIG_COMMIT_REV_ID: 88f73c1ee103bec7cd9dd7af4df3265dec0be562