[KGEN] Disable dead argument elimination pass for 24.4 (#40824)
The pass is miscompiling code when there is sufficient indirection
around function pointers. Since 24.4 is tomorrow, I'm disabling the pass
due to the severity of the issue and I have added a test case that fails
when it is on.

MODULAR_ORIG_COMMIT_REV_ID: fb2bbfa1757138d55d4502b631bc80227f10e99d