[Kernels] Remove overpadding of K when packing B (#40682)
The B tile needs to fit in the half the L2 cache in order to get
performance with matmul. This means that if K (MxNxK matmul) is too
large we have to run over blocks of K.

For example with AVX2 and a 512 KiB L2 cache this means the if K is
large than 2730 we have to run over blocks with K=2730. Currently we
over pad K so if K = 2731 we would pad to 2*2730=5460. This is a lot of
wasted space for one extra K value.

This PR no longer over-pads K (in some cases with in8t it might add at
most 7 bytes of extra padding).

Resolves KERN-468

LINALG_ORIG_REV_ID: b9b0e06221a4ae93f2c46d4fcf3ca1e5d5a4d712