[mojo] Remove `TaskGroupTaskList` and make tasks owned by the `TaskGroup` (#40847)
This PR resolves a landstanding TODO with the TaskGroup implementation:
it doesn't own the memory for its tasks. This required users to declare
a TaskGroupTaskList and manually keep the tasks in the task group alive
by extending the lifetime of the TaskGroupTaskList.

This was for 2 reasons: we didn't have a way to type-erase Coroutines,
but more importantly, destructors weren't a thing.

This PR puts the list of tasks, which the task group owns, inside the
TaskGroup. It gets type-erased because the underlying MLIR type is
opaque. The TaskGroup now owns the tasks and is responsible for
destroying it. It now returns a reference to the enqueued task.

ALGORITHM_ORIG_REV_ID: b66e579737be6c943e57d5d88379ce7f8a76c7a7