[MAX Driver] Allow MAX Driver to be used for KV cache (#40857)
@AustinDoolittle ran into some issues attempting to use the MAX Driver
APIs in his running KV cache example:
1. DeviceMemory and Tensor cannot be stored in a List because they are
not copyable
2. The device memory needed for the KV cache is not always a dense
tensor
3. TensorSlice is not fully flushed out for use in kernels yet

This PR attempts to workaround these issues by:
- Adding `__copyinit__` to DeviceMemory and Tensor so they can be stored
in a list, but disable the method add compile time using `constrained`.
This is a stopgap until MSTDL-467 is fixed.
- Adding a DeviceMemory ctor that takes a bytecount.

For point 3, we can temporarily create an NDBuffer wrapping the
Tensor._ptr, as long as we are careful to keep the Tensor alive while
the NDBuffer is in use.

Co-authored-by: Ian Tramble <ian@modular.com>
MOJO_MAX_ORIG_REV_ID: e00182baa3463e3a0c9395060e20e22819795020