[mojo-lang] Add LifetimeSetUnionAttr (#41011)
This attribute converts from a LifetimeSetType to a LifetimeType with a
given mutability. It does this by folding a LifetimeSetAttr through a
LifetimeUnionAttr. This is useful, because we want to make
`__lifetime_of` return a LifetimeSetAttr but also make it usable as a
regular, singular lifetime.

This PR adds the attribute definition and bytecode encoding.

MODULAR_ORIG_COMMIT_REV_ID: 16882a454e3b3056cd8a5f3dd7ec38f1ae1bc013