[mojo-lang] Lower Set Result/Error Slot (#40925)
This patch adds the error and result by ref arguments to the
continuation (Note that they are excluded in the frame). In summary,
this patch does the following:

- omits error values and result values from frame
- adds special case to extract error and result values directly from
continuation in the FrameVariables cache
- lowers the SetError/Result CO Op to the proper store into the frame
- transforms coroutine types into header types.
 This is necessary for populating the error/result slot
when we don't know the full type of the coroutine.

MODULAR_ORIG_COMMIT_REV_ID: 96ef9f33908b696d9f9bed3c5a38480a6560d9a1