[KGEN] Add "typeValue" field to TypeConstantAttr (#40997)
Project: [MOCO-739 : \[KGEN\] Value-Domain
[Internal link]

Add "typeValue" field to TypeConstantAttr. It is meant to carry the
value-representation of a type.

A temporary builder that only takes in one `mlirType` and duplicates it
for the `typeValue` is created for back-compat. It's meant to be very
short-lived as that kind of usage is very likely to be an unintentional
mistake in the future when the type & value domains are properly split
up (That will come when followup PRs introduce `TypeValueType` and
`SourceStructType`).

The printer/parser will ensure that when the two reprs are the same, it
is only printed once, so no existing IR needs to change for the simple
case.

MODULAR_ORIG_COMMIT_REV_ID: 9791d3ccd1576aeea59be31876b403b3c5434cf1