[mojo-lang] Implement named results with a decorator (#41041)
This PR implements named result slots. This is a feature where the
`byref_result` argument of a function that has an in-memory result is
mapped to a decl, so that it can be used directly to emplace values.
This PR implements it with a `@__named_result(out)` decorator to bypass
any syntax concerns. I put an item on next week's deep dive to bikeshed
this syntax.

This decorator doesn't work for functions that don't nominally have an
in-memory result. We should probably change it so that if used with a
register-passable result, it generates an `byref_result` result anyways.

MODULAR_ORIG_COMMIT_REV_ID: 9a67a52a5e84d7e1b467c3b1806f3923a250c358