[KGEN] Teach LowerClosures to handle byref results in async functions (#41085)
Add support to the legacy path so we can get this working in the live
path, even if it's going to be inefficient. We'll circle back when the
new lowering is live and focus on performance. E.g., we geerate the
stores and loads of the error slot even if it's not needed (we also
should not generate it if it's not needed).

MODULAR_ORIG_COMMIT_REV_ID: a6cb8e1c12dcecaf9b9f9f34f6b35a4b9fc7ed9b