[KGEN] Teach `co.execute` about byref results (#41086)
This PR changes the spec of `co.execute` to allow it to have up to 2
arguments to its body region, representing the `byref_error` and
`byref_result` arguments of the function respectively. These arguments
are needed because the result slots are not set at the callsite of the
async function/region, but when it is awaited.

Incidentally, this "just works" with the current inlining code, so just
add a test to make sure async functions with byref results can be
inlined. Also teach LowerClosures how to handle these in
lowerAsyncExecute in a similar way to lowerAsyncFunction.

This should provide the necessary scaffolding to support byref results
in async functions in the frontend.

MODULAR_ORIG_COMMIT_REV_ID: a5e87706f1f8feca80723c03e69f462d59d37a68