[mojo-stdlib] Move global UnsafePointer functions to be methods.
This extends the compiler to allow the type of `self` to be `Self` with
any parameters, allowing more flexible conditional conformance.

This allows us to move the global functions in UnsafePointer to being
methods on it instead, which is a much nicer API.  Specifically, this
moves:
 1) `destroy_pointee(p)` => `p.destroy_pointee()`
 2) `move_from_pointee(p)` => `p.take_pointee()`
 3) `initialize_pointee_move(p, value)` => `p.init_pointee_move(value)`
 4) `initialize_pointee_copy(p, value)` => `p.init_pointee_copy(value)`
 5) `move_pointee(src=p1, dst=p2)` => `p.move_pointee_into(p2)`

This uses the word "init" instead of "initialize" since that is the term
of
art used in Python.

MODULAR_ORIG_COMMIT_REV_ID: ba930ea1ef25c4f010bd27742e3afee047377764