[mojo-lang] Add `__get_nearest_error_slot` magic function (#41103)
This magic function returns an MLValue for the nearest error slot,
whether `byref_error` function argument or `lit.try` vardecl for the
error. This allows the error of the surrouding context to be directly
emplaced. Importantly, it allows the error pointer to be passed into a
RaisingCoroutine so that it can be retrieved when the error function
executes. This will allow raising tasks to store an error.

MODULAR_ORIG_COMMIT_REV_ID: 8165e9c89489c00eeaa7920c4805747fcda133c5