[KGEN] Tiny optimization for async functions that return `None` (#41242)
We know that storing a `none` value to a pointer is a noop, and an
`alloca` of a none element returns a nothing pointer. We can use this to
elide storing the result pointer of an async function with a none
result (and also a load if it was loaded from somewhere else).

This removes one load and one store for each call to a none async
function, which calms the optimization demons in my head.

MODULAR_ORIG_COMMIT_REV_ID: b0c221a1326ea4ac61887d4fdc9ed8f294e9555c