[KGEN] Factor call op inlining behind KGENCallOpInterface (NFC) (#41271)
This is a change I've been meaning to do for a long time: have
KGENCallOpInterface define the interface for inlining a call operation.
I finally decided to do it in this PR because I want to cut the
dependency on CODialect from TransformUtils, so that CODialect can
dependend on TransformUtils for inlining.

MODULAR_ORIG_COMMIT_REV_ID: f8bbd0f2123a4944c78c09b7428d91775c4fa816