[KGEN] Implement `co.await(co.execute) -> inline` optimization (#41346)
Semantically, `co.await` of a coroutine calls into the resume function
and awaits for it to complete. This means given the execute region, we
can inline the call.

This is one of the major optimizations: when an async function call of
the form `await foo()` is inlined, the call overhead gets optimized
away.

MODULAR_ORIG_COMMIT_REV_ID: 8b6e1f6270a90679f008c3a16dad1324b3a6e947