[MAX Driver] Use DeviceContext APIs in MAX Driver (#41309)
This removes all uses of CUDASupport in the MAX Driver impl, and instead
interacts with CUDA through the DeviceContext APIs in mojo.

Since the DeviceContext is written in mojo, and the core of the MAX
Driver is (currently) in C++, we need to pass a struct of function
pointers from mojo to C++ for the C++ side to have access to these APIs.
Eventually it might make sense to rewrite more the MAX Driver in mojo to
avoid these shim layers, but this works for now, and is convenient for
the graph/engine APIs since they are mostly in C++.

Fixes MSDK-377.

---------

Co-authored-by: Ian Tramble <ian@modular.com>
GPU_ORIG_REV_ID: 806fdc1c7004f13df236b2a20c5a8cdc0960aed8