[Support][NFC] Move Atomics from LLCL to Support (#41479)
### [Support][NFC] Move Atomics from LLCL to Support

This break the dependency that `MeteringContext` has on `LLCL/`, so that
we can eventually make avoind any dependency of `Support/` on `LLCL/`.

This change moves `Atomics` into `Support/Threading` and updates all of
the code which references `Atomics` to use it in the new location.

Towards GRA-595

MODULAR_ORIG_COMMIT_REV_ID: b0b63033095907ebb8acc88a824092145161debb