[Tests] Disable telemetry when invoking the test executor (#41429)
The test executor shouldn't be blocked by any kind of telemetry, as it
should be as fast as possible, so it's better to disable telemetry. We
can do that easily via a env var.
On the other hand, `mojo test`, which is not the actual executor, has
telemetry enabled, which is what we want.

MODULAR_ORIG_COMMIT_REV_ID: d32c21b1927b0a7fc53190c5c93e409f81f196f7