[DebugInfo] Fix scope of subprogram (#41677)
Fixes [MOCO-826 : \[mojo-parser\]\[debuginfo\] Wrong subprogram scope
for imported
[Internal link]
Have the generated scope of subprograms just be the file scope. This is
not correct for nested functions but this is no different from the
previous behavior, which at best would find a file scope from the scope
stack (and may not even be correct).

I'm planning on a larger refactor of how debug scopes are handled in the
parser. It seems a global scope stack may not be well suited for our
parser, since it doesn't always parse in a stack/tree manner (it can
jump to a sibling branch in the ASTDecl tree when resolving calls for
example). It may be easier to just track the debug scopes alongside the
AST/IR tree.

MODULAR_ORIG_COMMIT_REV_ID: 3138b40871cffae0841671cce896bf6e4ac73812