[Kernels] Q4_K refactoring (#41704)
This change refactors code related to Q4_K zero point handling. This
chunk of the kernel involves doing a mini-matmul of int16 factors. This
had been optimized for VNNI usage, but this is now cleaned up to use
NEON intrinsics too.

After a bunch of experimentation trying to decode the scale/min arrrays
inside the M=1 kernel, I gave up and will continue with decoding those
arrays to the stack and using them inside the M=1 kernel. This is
pushing more of the post processing in the current kernel to be moved
out so that it can be shared across the two variants of kernels.

QUANTIZATION_ORIG_REV_ID: 61c8fa8f0f5c4c4195c7110663525b9c91935753