[MGP][Runtime] Remove cuStream from MojoCallContext (#41762)
Small cleanup PR to remove the cuStream in MojoCallContext. We already
store the stream in the runtime which is passed to the call-context.

Runtime is kept for now, but it can be removed once we have explicit
device Context in IR.

MODULAR_ORIG_COMMIT_REV_ID: d32abf8e07a29073aba302745a7d4d55d5b97afe