[mojo-lang][CheckLifetimes] Fix parameter for + else handling. (#41812)
This revises CheckLifetime and LowerSemanticCF to properly model
`@parameter for`.  This for statement is not the same as an `hlcf.loop`:
`hlcf.loop` is an infinite loop with zero or more breaks in the body,
which is guaranteed to have an empty `else` by the time it gets to
CheckLifetimes.

In contrast, a `@parameter for` is a /finite/ loop that eventually falls
out
the bottom of itself (going into its attached `else` block which may be
non-empty).  Instead of using contortions to try to make one look sort
of like the other, switch to modeling it correctly.

This exposes other problems.  `hlcf.elif` isn't doing the right thing,
which I'll address in other patches.

MODULAR_ORIG_COMMIT_REV_ID: 5e472546af895ae1e98fce9cfdfcf4a01db09804