[mojo-lang][CheckLifetimes] Revise CheckLifetimes dtor insertion (#41838)
This is a substantial rework of a lot of recent progress in
CheckLifetimes, and
fixes bugs 'self' in initializers with exception handling. This fixes
MOCO-579
and Fixes MOCO-801.

Some more details:
1) Treats "self" as being fully initialized in an init when all the
fields are
initialized. This means that throwing after all the fields are set
causes
the dtor on self to be run, but setting some fields will cause just
those
   field destructors to run.
2) Rewrites the ElifOp handling logic (in both passes) to be a lot
simpler
    and use standard helpers instead of duplicating the logic.
3) Removes error prone logic that was trying to incorrectly infer what
the
'markInitialized' is doing in an EH region for a call, and removes logic
     that was special casing breaks in loop processing.
3) Ends up with significantly less code.

MODULAR_ORIG_COMMIT_REV_ID: c43268be9e5565e4a2fde04a2829cf9e7b099667