[mojo-lang] Chop Off From Back Not Front (#41722)
The code that mapped the async function arguments from the original to
the ramp function was incorrect. It was chopping off the first N
arguments where N in {0,1,2} depending on the in memory result and error
fields of the function signature. This patch corrects that logic so it
chops off the last N arguments instead.

MODULAR_ORIG_COMMIT_REV_ID: 4413d841aef1f3266d1d76d9338f6d8e094e3d36