[stdlib] Fix `libm` detection in `math` for floats
Detecting whether `libm` is supported for a given type is incorrect.
Previously, it checked if the type was both float32 *and* float64, which,
of course, is impossible to be true.  So, we were only saying `libm`
was supported for integral types.  This is incorrect.  Fix this
to properly use *or* so that we claim libm functions are supported
for both float32 and float64 types.

MODULAR_ORIG_COMMIT_REV_ID: d579662bfb2d3550dab9355697847b27b93b3fd6