[MAX Driver] Rename AnyMemory->AnyTensor (#41915)
This is meant to be a convenience wrapper to represent Tensors with
different compile-time rank and dtypes. "Memory" has a runtime size
and no compile time values associated with it so does not make sense
to use in this context.

MOJO_MAX_ORIG_REV_ID: c2571a5b03103591110fa6b3b3deea07334d245f