[mojo-lang] Wrap throwing finally regions in try-except (#41998)
The finally block executes whenever control flow leaves any of the other
regions of a `try`, whether through a yield, return, or raise. Its
overall
control flow effect takes precedence over how control flow left the
other
`try` regions originally.

This is conceptually implemented by branching before said exits to the
finally region, and when yielding from the finally region, branch back
to
where control was before.

However, this means that the finally region can conditionally overwrite
an
error slot and then choose not to raise, causing this issue:

```
def raising_finally():
    try:
        raise Error() # initializes %__error__, then branch to 'finally'
    finally:
        # conservatively destroy %__error__ before the raising call
        might_raise()
        # if the call didn't raise, branch back to where we were in 'try',
        # but now we have an error return with an uninitialized %__error__!
```

Thus, the error slot in the finally region cannot alias the one used in
the
other regions, because it might conditionally overwrite it while it
still
needs to be used. This PR fixes this by rewriting the above into:

```
def raising_finally():
    try:
        raise Error()
    finally:
        try:
            might_raise()
        except:
            raise
```

This will cause the error used by the finally region to get emitted into
a temporary, move into the overall error if we know for sure it emitted
an error. The same is done for `with` statements, since `__exit__` is
allowed to raise, apparently.

MODULAR_ORIG_COMMIT_REV_ID: 15705578bb65fb0987677c1e7726ee14aaef86e0