[MGP] Make `generic.compiled_region` take in chains and output chains explicitly in operands/results (#41765)
Current mode of operation is that we implicitly assume during the `kgen`
compilation that we don't have any chain structures and we will not have
to lower them. They are removed post seq.fusion. As a result any kernel
tensor or opaque type does special handling for the chains. This is to
remove that implicit removal, and have input and output chains explicit
in the generic operands/results.

This PR removes the special pass that removes chains + adds lowering for
the chains in both kernels and the primitives pathway. lower-mogg-abi
will handle the chain outputs and generate them at the boundary of the
generic region as part of result list.

Chain representation in IR of generic region. It is generically handled
as an operand/result
```mlir
    %2:3 = mogg.generic.compiled.region (%arg0, %1) : (%arg2: !mef.chain, %arg3: index) -> (!mef.chain, index, index){
      %5:5 = mogg.get_async_vals() : () -> (!kgen.pointer<none>, !kgen.pointer<none>, !kgen.pointer<none>, !kgen.pointer<none>, !kgen.pointer<none>)
      %6 = mogg.call "builtin.unpack_async"(%5#1) {} : !kgen.pointer<none> -> !kgen.pointer<none>
      %7 = pop.pointer.bitcast %6 : !kgen.pointer<none> to !kgen.pointer<index>
      %8 = pop.load %7 : !kgen.pointer<index>
      %9 = mogg.rt_placeholder : !kgen.pointer<none>
      %10 = "mogg.create_ready_chain"() : () -> !mef.chain
      %11 = mogg.call "mip.constant.index"() {value = 42 : si64} -> index
      %12 = mogg.call "mip.constant.index"() {value = -20 : si64} -> index
      %13 = mogg.call "mip.add"(%11, %12) {} : index, index -> index
      %14 = mogg.call "mip.add"(%13, %8) {} : index, index -> index
      %15 = mogg.call "mip.mul"(%14, %13) {} : index, index -> index
      mogg.call "builtin.create_index_async"(%11, %5#3, %9) {} : index, !kgen.pointer<none>, !kgen.pointer<none>
      mogg.call "builtin.create_index_async"(%15, %5#4, %9) {} : index, !kgen.pointer<none>, !kgen.pointer<none>
      mogg.call "builtin.create_chain_async"(%5#2, %9) : !kgen.pointer<none>, !kgen.pointer<none>
      mogg.output %10, %11, %15 : !mef.chain, index, index
    } {canDuplicate = true, canFuseIntoOther = true, hasCuda = false, inOutSignature = #M.inout_sig<"">, isAsync = false, summary = "mogg.generic.compiled.region:30+mogg.generic.compiled.region:34+mogg.generic.compiled.region:38+mogg.generic.compiled.region:44+mogg.generic.compiled.region:48+mogg.generic.compiled.region:52"}
```

Two follow-ups need to happen after this lands:
* Remove special output and input chains for the region should be
removed in favour of the ones in operand/result list
* Elide all the chain-outputs when fusing - since chain primitives are
still in C++, we don't have a lot of async chains so this can come
after.

MODULAR_ORIG_COMMIT_REV_ID: a1bab6c4d736d3411616f632bffd4b7f9d006792