[mojo-lang] Fix try-finally regions in not raising contexts (#42035)
A try-finally block in a non-raising context would silently ignore any
raised errors, causing a huge safety issue.

```mojo
fn cannot_raise():
    try:
        raise Error()
    finally:
        print("hi")
```

This function would swallow the error and continue execution. The
default python behaviour of a try-finally block is to re-raise the
error. In our case, we want to make the 'try' block a non-raising
context if its outer context is non-raising AND it does not define an
'except' region to explicitly handle the erro.

To do this, when parsing a 'try' block in a non-raising context, we need
to lookahead to see if an explicit 'except' block is defined. If not, we
have to emit a `kgen.unreachable` into the except region.

MODULAR_ORIG_COMMIT_REV_ID: 44f0161403e232a3869b57e4a9fe034b58ccd147