[bazel] Fix rpaths with asan (#42073)
In some cases mojo dlopens libraries that then dlopens other libraries.
This fails when running with asan because of this issue:

https://github.com/llvm/llvm-project/issues/28164#issuecomment-2181788143

To fix this we can add the rpaths to mojo itself to potentially catch
this. This isn't ideal but should be harmless.

This rewrites the rpath_shim rule to use string paths vs hard
dependencies to avoid circular dependencies and unnecessary building.
The only downside to this is if we move things it won't immediately fail
even though the label no longer exists. That should be minor since we
will catch them in the tests that were failing anyways.

MODULAR_ORIG_COMMIT_REV_ID: b43b57f1b738ed5a7e581e8c2011b1ac4b490c19