[Python API][Support] Remove MPython library. (#42153)
The `MPython` support library turned out to be an issue due to requiring
pybind and python, but we build the python bindings separately for each
python version. Both the torch-mlir python bindings and the core python
API require the helpers in `MPython` though, so for now the easiest way
to avoid problems is to duplicate the helper implementations for each.

MODULAR_ORIG_COMMIT_REV_ID: a91b2969f58335fc0869617b16fa12af502978aa