[Kernels] change qmatmul_k to use streaming matmul interface (#42143)
Before this PR, the low level matmul loops for x86/NEON would directly
read from the unpacked 4/6 bit values on the stack. For the general
case, unpacking the 4/6 bit weights saves compute time by avoiding
repeated bit shifts/masking.

With this PR, the low level matmul loops take a lambda to stream in the
weights. With the existing kernels, that lambda reads from the stack
buffer and this is effectively NFC.

When the M=1 code lands, there is an alternate lambda that reads the
quantized weights instead and unpacks on the fly. Unpacking to the stack
and reading back from the stack with no data reuse is a performance
loss.

QUANTIZATION_ORIG_REV_ID: b2ce019a2d48174b924c196d27d5ae85d8cd0d9c