[KGEN] Legalize against materialization of capturing parameter closures (#42186)
Fixes
[Internal link]

Parameter closures that are capturing cannot be materialized into a
runtime value. Previously this was only checked at the top-level for
ParamConstantOp, but we also need to check if any part of a param
constant is a capturing parameter closure, as the closure could be a
field in a struct for example. This can be easily checked at LLVM
lowering time where we recursively materialize param constants.

Ideally we warn the user at the parser level too (today we're also [only
checking if the top level type violates
[Internal link]
but I'm not sure it's worth duplicating the materialization logic there
or not 🤔, since we need special type-walking logic to look for it.

MODULAR_ORIG_COMMIT_REV_ID: f6f80eb81d2ec1038a390b6970d5eee9a30ff0e9