[Support][Docs] Add a README for `Support/` (#42218)
This is the first step towards completing the documentation of the
`Support/` directory. It adds a README file, indicating the purpose of
the directory, plus the build and the documentation requirements for
libraries in this directory.

I'll follow up with additional PRs to add documentation for each
library.

Towards GRA-595

MODULAR_ORIG_COMMIT_REV_ID: 79099e29d351ddb08c74acd603d5c9645ca929e0