[Kernels][GPU] Enable kernel profiling via DeviceContext. (#42211)
- Adds kernel timing info collection via DeviceContext upon enqueuing
kernel (with appropriate flag). Collected in a list of structs that can
be extended to include more info down the line.
- Adds functions to print kernel timing information and clear the list.
- Adds an example test (`test_deviceContext_profiling.mojo`)

Is a remake of [Internal link] in that:
- we use a separate struct to record the kernel profiling information
and pass the pointer to a list (kernel profiling info is NOT stored in
`DeviceContext` itself; this simplifies code - avoiding requirement of
`inout` throughout the code - and redicing the number of constraints on
the context itself).
- we use a struct (vs. dictionary before) that can be extended if need
be to include more profile info.

Closes KERN-584

GPU_ORIG_REV_ID: ecb1d92f5d22b754af9b6ae7aa63874dca3e511d