[PT] Restructure TorchMLIR Python bindings linkage and add support for loading torch MlirModule with InferenceSession (#42155)
In order to pass `MlirModule` objects created with the Python bindings
we need to load the MLIR C API symbols from the same source as `mtorch`
to maintain cohesive `TypeID` values of MLIR objects. To achieve this
the C API is linked in to the `mtorch` library which is then used both
by the torch-mlir Python bindings and `mtorch`.

This is a step towards using the fx importer for lowering
`torch.fx.Graph` objects in Python to a `MlirModule` with the Torch
dialect. With this change it is possible to create a custom backend for
`torch.compile` that lowers a `fx.Graph` and compiles and loads the
resulting `MlirModule` with `InferenceSession`.

MAX_PYTHON_ORIG_REV_ID: e01b8fff06493819ee4811042bff553dc631362d