[mojo-lang] Support raising with statements with unconditional exits
When using a context manager in a raising region, the language requires
the context mangaer to define `fn __exit__(self, e: Error) -> Bool`.
This method conditionally handles errors in the `with` region, allowing
it to be swallowed instead of propagated.

This is a useful feature for some applications, but in most cases, users
just want the context manager to unconditionally exit: many of these are
defined as trivially delegating to the unconditional exit. However,
because the compiler still things that the `with` may conditionally
exit, it still treats code as reachable along a raising path. For
example, the following now work:

```mojo
def foo() -> Int:
    with ContextMgr():
        return 2
    # no longer complains about missing return

def bar():
    var x: Int
    with ContextMgr():
        x = might_raise()
    use(x) # no longer complains about uninitialized
```

This is because the compiler statically knows that if the `with` region
raises, the context manager always allows the error to propagate.

MODULAR_ORIG_COMMIT_REV_ID: a4cc4af3d5d3af36094e9d74b4a177505a2805ef