[bazel] Remove unnecessary CompilerRT dependencies (#42281)
This makes it more clear that we're not relying on the CompilerRT that
was linked with the most binaries. I don't think this was ever
happening, but I was worried it was. This revealed we were incorrectly
depending on CompilerRT in rules because we weren't propagating its
runfiles. This previously wasn't a problem because its runfiles were
also part of the host tool's runfiles, but that isn't the case now for
`kgen` which doesn't have a dependency on CUDASupportGlobals.

MODULAR_ORIG_COMMIT_REV_ID: 981b28e533386fddb81c59bf9bb7a314b43b9893