[KGEN] Fix compile assembly stub function debuginfo (#42422)
Fixes [Internal link]

The root cause is when we have a function with a parameterized signature
(and hence a parameterized DISubroutineType) get used by
`compile_assembly`. When creating the asm_stub for this function, we
have the concretized signature type (from the compile_assmebly call),
but we don't have the concretized DISubroutineType to assign the newly
created asm_stub. This becomes difficult if the parameter in the
DISubroutineType was a lifted param in the body of the original
function.

This PR re-creates a new SubroutineType based on the concretized types
from the call signature. It's still not perfect though, since
technically the stub is a compiler-synthesized function, and should not
have a debug frame (but it has to have a debug frame because it's the
root frame), but this is a much lower priority problem.

MODULAR_ORIG_COMMIT_REV_ID: b2c00f5ab331bd356f394474cac31c88e7283847