[KGEN][MOGG] Add reduced version of auto-param (#40446)
Adds a pass which adds a TensorSpec parameter to every kernel which
takes a tensor argument and propagates those through the call stack.
Starting from the kernels it traverses the call graph out to identify
the leaves which actually need the specs and are reachable from the
kernel. It then adds the parameters to those nodes and rewrites the
stack all the way back to the kernel

This is a much reduced version of the version in #34958

MODULAR_ORIG_COMMIT_REV_ID: 106719c97e5cfdc385f0d2b7d5dedca999528761