[Support] Make ASSERT_STREAM message more general (#42409)
Since `ASSERT_STREAM` is now part of the shared `Support` library, it
should not have a message specific for any tool or executable.

This change removes the tool-specific message for MAX (it will re-appear
in a later PR, specific to the graph compiler tools). The change also
makes the assert message more similar to that of `assert`, with the hope
that developers will be able to easily parse it visually.

Fixes GRA-666

MODULAR_ORIG_COMMIT_REV_ID: 54f15906aac060983f38bf37901ffc8aa2fae242