[KGEN] Tell llvm that mutable arguments are noalias (#42715)
This is the big one. One of the big guarantees of the language is that
each function can only have one mutable reference to a value at a time.
This means that any mutable in-memory arguments are known to be noalias.
The only missing thing is that we don't know if the `ref` argument
convention corresponds to a mutable or immutable lifetime after
elaboration, when it gets resolved. We should find a way to propagate
ths information, and aliasing information implied by lifetimes in
general, to LLVM.

MODULAR_ORIG_COMMIT_REV_ID: 242f5c089f02641b5b2a467f98b072040bc6e063