[stdlib] Use sizeof[DType] and bitwidthof[DType] when DType is known at compile time
This replaces `DType.sizeof()` with `sizeof[DType]()` and
`DType.bitwidth` with `bitwidth[DType]()` when `DType` is a parameter.
This is cleaner because DType is treated as a parameter and is
guaranteed to work with target-dependent types.

MOJO_MAX_ORIG_REV_ID: 2ab52a594fa1ca1e525f8cc50360284c56b6b42a