[KGEN] Add symbol to `kgen.custom.op_impls` for faster access (#42789)
Adding a symbol allows to quickly retrieve the operation, compared to
walking the module block.

The operation is defined as exported so the operation is not DCE'd.

One test had to be changed as there was a different error when there was
no symbols, or if there was no `main` symbol. This PR unifies both
errors.

MODULAR_ORIG_COMMIT_REV_ID: 34f1260f5bc0af461aad2f555ad0c2101be56c0e