[mojo-stdlib] Mostly remove `*Pointer(address=` constructors (NFC)
This PR deletes the `address=` constructors for UnsafePointer,
LegacyPointer, and removes all but one use of it for DTypePointer. It
also removes the constructor from `Scalar[DType.address]` to
LegacyPointer, but users of the constructor for DTypePointer remain in
the masked gather and scatter intrinsics. This is part of removing the
ability to convert integers to pointers, in an attempt to move away from
C pointer semantics.

MODULAR_ORIG_COMMIT_REV_ID: 68841ea1e4ff262abe1f507d087fb29b55deb577