[stdlib] Fix UInt strided range length
The `len` of a `_UIntStridedRangeIterator` was incorrect in the case where the
`end > start` for a UInt strided range.  Specifically, we would do `start - end`
(which would be a negative result), but the return type is a `UInt`, so we would
end up reporting the length as a very large `UInt` number.  This would manifest
when iterating UInt strided ranges since we would iterate *way* too many times.

Fix this behavior by just returning `0` for the `len` in this case.

MODULAR_ORIG_COMMIT_REV_ID: c467e1b960c9bbe253584db5f787620af1c036a4