[mojo-stdlib] Remove `DType.address`
`DType.address` is an alternate representation of pointers in the SIMD
dtype system. Its use cases are extremely narrow and we don't need or
want this generality in the representation of the language. Remove it
entirely and zoom in on the two use cases the matter: masked scatter and
gather (interop with LLVM intrinsics).

MODULAR_ORIG_COMMIT_REV_ID: d14cfa2460a44280288883687497e8cbe12ce312