[mojo-lang] Move determinism.test to mojo-integration target (#42982)
This takes a really long time to run because kgen-translate is
running 4 times serially, parsing the stdlib in a non-isolated
way.  This takes forever (and will likely to continue to get
worse as the stdlib grows) so move it to mojo-integration to
avoid impacting core dev workflows.

It would still be great to isolate this, but I don't want to
lose the value the test is providing.

MODULAR_ORIG_COMMIT_REV_ID: 977c9ef4f37f94375289d46ec6925ef525ce963a