[mojo-lang] Push CachedTypeLifetimeFinder through the stack. (#42981)
This is a follow-on to #42005 which puts the cache in MojoParser
SharedState (so it is a more useful cache) and updates
`getOperationEffects` to use it (likely a correctness fix).  The
major point here is to be a (likely inconsequential) compile time
perf win now, but that also allows us to work more generally with
lifetimes in the future without regressing compile time.

This is ~NFC other than the likely correctness fix, but the tests
will come when we start enforcing exclusivity.

MODULAR_ORIG_COMMIT_REV_ID: eccf2162df27e85bdcc47f56f5e95dee1655e9cf