[mojo-lang] Rename `DeclRefType` => `StructType`. NFC (#42989)
This renames the internal DeclRefType to StructType since this is how
we form a reference to a struct, and this is more consistent with
TraitType.  This is done after discussion with Jeff offline.

This annoyingly conflicts with the KGEN StructType a lot, requiring a
lot of LIT:: qualifiers similar to what we have for FuncOp. I think that
a good solution for this is to rename the KGEN type to "RecordType"
instead of "StructType" but won't do that unilaterally.

MODULAR_ORIG_COMMIT_REV_ID: 3334a910126424ad43edc178aa1353b51610b0c8