[Kernels][mojo-stdlib] Get `Dim` off of `OptionalReg`. (#42986)
This replaces the implementation of Dim to just use a sentinel in
the Int value instead of using OptionalReg (which we'd like to
remove).

Note that `Dim` is extremely unsafe, it's get method would
previously just return an incorrect value out of the variant when
accessed and not-present.  This maintains that behavior, but it
makes me very nervous, and should probably be an assert.

GPU_TEST_ORIG_REV_ID: 88e32666faaf605c30f6a20b3c1327b49bc2e969