[mojo-lang] Fix longstanding problem resolving static methods. (#42991)
Mojo allows you to call a static method on a struct with either
`T.method()` or `someT.method()` where the later is supposed to
drop the value (like Python does) when the method doesn't take it.

We had a long standing problem that prevented us from overloading
methods on static-ness, which was due to how and when we resolved
this bit.  Overload resolution was mostly ready for this, but we
didn't propagate it out of overload resolution.

Fix this by having resolveOverloadSet mutate the operand list when
the self operand is being intentionally ignored, making it so
call emission then does this right thing.

This resolves a long standing fixme in `AttributeRefNode::emitIR`.

MODULAR_ORIG_COMMIT_REV_ID: 0000719063304ddb7a3c75b93700eb43be603f1d